jobbotron part 5 complete

+++ w6/express-pg/routes/jobs.js
@@ -3,14 +3,20 @@ const router = express.Router();
 const db = require('../db/index');
 const bcrypt = require('bcrypt');
 const jsonwebtoken = require('jsonwebtoken');
-const { ensureLoggedIn, ensureCorrectUser } = require('../middleware/auth.js');
+const {
+  ensureLoggedIn,
+  ensureCompanyAcct,
+  ensureCorrectCompany
+} = require('../middleware/auth.js');
 
 // POST /jobs
-router.post('', async function(req, res, next) {
+router.post('', ensureCompanyAcct, async (req, res, next) => {
   try {
+    const token = req.headers.authorization;
+    const decodedToken = jsonwebtoken.verify(token, 'CONTIGO');
     const data = await db.query(
       'INSERT INTO jobs (title, salary, equity, company_id) VALUES ($1, $2, $3, $4) RETURNING *',
-      [req.body.title, req.body.salary, req.body.equity, req.body.company_id]
+      [req.body.title, req.body.salary, req.body.equity, decodedToken.id]
     );
     return res.json(data.rows[0]);
   } catch (err) {
@@ -41,15 +47,24 @@ router.get('/:id', ensureLoggedIn, async (req, res, next) => {
 });
 
 // PATCH /jobs/:id
-router.patch('/:id', async function(req, res, next) {
+router.patch('/:id', async (req, res, next) => {
   try {
+    const currentCompany = await db.query(
+      'SELECT company_id FROM jobs WHERE id=$1',
+      [req.params.id]
+    );
+    const token = req.headers.authorization;
+    const decodedToken = jsonwebtoken.verify(token, 'CONTIGO');
+    if (currentCompany.rows[0].company_id !== decodedToken.id) {
+      return res.json({ message: 'Unauthorized -- wrong company' });
+    }
     const data = await db.query(
       'UPDATE jobs SET title=$1, salary=$2, equity=$3, company_id=$4 WHERE id=$5 RETURNING *',
       [
         req.body.title,
         req.body.salary,
         req.body.equity,
-        req.body.company_id,
+        decodedToken.id,
         req.params.id
       ]
     );
@@ -60,8 +75,17 @@ router.patch('/:id', async function(req, res, next) {
 });
 
 // DELETE /jobs/:id
-router.delete('/:id', async function(req, res, next) {
+router.delete('/:id', async (req, res, next) => {
   try {
+    const currentCompany = await db.query(
+      'SELECT company_id FROM jobs WHERE id=$1',
+      [req.params.id]
+    );
+    const token = req.headers.authorization;
+    const decodedToken = jsonwebtoken.verify(token, 'CONTIGO');
+    if (currentCompany.rows[0].company_id !== decodedToken.id) {
+      return res.json({ message: 'Unauthorized -- wrong company' });
+    }
     await db.query('DELETE FROM jobs WHERE id=$1', [req.params.id]);
     return res.json({ message: 'Job deleted' });
   } catch (err) {
